Add opt-in support for private git config
This update removes the user.name, user.email, etc from .gitconfig
and adds a hook so that, if a file `~/.gitconfig.local` exists,
it will be read by git.  This allows for you to set private git
configurations that are not committed into source control but are
still supported by the dotfiles and recognized by git.